+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `mirror_leg` adds the ability to mirror a leg instantly.
+- `add_opposite_leg` adds the ability to mirror a leg instantly.
 - Using `uv` instead of raw Python.
 
 ### Changed